requirements.txt
blosc
u-msgpack-python
